Rename shared tRPC React API file
